C code for MSP430G2553
I have added the function of music playing to the MCU system in replace
of the boring buzzi. To accomplish this, I have to use a timer to
generate PWM pulse output to produce tunes with different frequencies by
which it can play music. Since there are only two timers on MSP430G2553
with PWM function and they are used for timing before, I replace one of
them with Watch Dog Timer which also provide a timing function and  but
not PWM. The timer spared is used for playing music.